Backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==1.7.1
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing-extensions==4.7.1
